refactor: delete genres reducer, change movies reducer
