--- src/modules/genres/genresSelector.js
@@ -1,3 +1,0 @@
-export const getGenres = (state) => state.genres.genres;
-export const getGenresIsLoading = (state) => state.genres.isLoading;
-export const getGenresError = (state) => state.genres.error;
