+++ src/modules/movies/moviesAction.js
@@ -7,6 +7,9 @@ export const FETCH_MOVIES_SUCCESS = 'FETCH_MOVIES_SUCCESS';
 export const FETCH_MOVIES_ERROR = 'FETCH_MOVIES_ERROR';
 export const SET_MOVIES_CONDITION = 'SET_MOVIES_CONDITION';
 export const SET_SEARCH_QUERY = 'SET_SEARCH_QUERY';
+export const FETCH_GENRES_IS_LOADING = 'FETCH_GENRES_IS_LOADING';
+export const FETCH_GENRES_SUCCESS = 'FETCH_GENRES_SUCCESS';
+export const FETCH_GENRES_ERROR = 'FETCH_GENRES_ERROR';
 
 export const fetchMoviesIsLoading = () => ({
   type: FETCH_MOVIES_IS_LOADING,
@@ -27,7 +30,12 @@ export const setMoviesCondition = (condition) => ({
   payload: condition,
 });
 
-export function fetchMovies(condition, genres, query) {
+export const fetchGenresSuccess = (content) => ({
+  type: FETCH_GENRES_SUCCESS,
+  payload: content,
+});
+
+export function fetchMovies(condition, query) {
   return (dispatch) => {
     dispatch(fetchMoviesIsLoading());
 
@@ -56,13 +64,18 @@ export function fetchMovies(condition, genres, query) {
         break;
     }
 
-    return fetchQuery(genres, query)
-      .then((res) => {
-        dispatch(fetchMoviesSuccess(res));
-        return res;
+    return movieDbApi.getGenres()
+      .then((genres) => {
+        dispatch(fetchGenresSuccess(genres));
+        return genres;
       })
-      .catch((error) => {
-        dispatch(fetchMoviesError(error));
-      });
+      .then((genres) => fetchQuery(genres, query)
+        .then((res) => {
+          dispatch(fetchMoviesSuccess(res));
+          return res;
+        })
+        .catch((error) => {
+          dispatch(fetchMoviesError(error));
+        }));
   };
 }
