+++ src/modules/movies/moviesSelector.js
@@ -2,3 +2,4 @@ export const getMovies = (state) => state.movies.movies;
 export const getMoviesIsLoading = (state) => state.movies.isLoading;
 export const getMoviesError = (state) => state.movies.error;
 export const getMoviesCondition = (state) => state.movies.condition;
+export const getGenres = (state) => state.movies.genres;
