+++ src/modules/store.js
@@ -3,12 +3,10 @@ import {
 } from 'redux';
 import thunk from 'redux-thunk';
 import moviesReducer from './movies/moviesReducer';
-import genresReducer from './genres/genresReducer';
 import trailerReducer from './trailer/trailerReducer';
 
 const rootReducer = combineReducers({
   movies: moviesReducer,
-  genres: genresReducer,
   trailer: trailerReducer,
 });
